better labels in resource tables
